+++ spec/features/contextual_navigation_spec.rb
@@ -26,6 +26,7 @@ describe "Contextual navigation" do
     and_i_visit_that_page
     and_i_see_the_transition_contextual_breadcrumbs
     and_i_see_the_transition_related_links
+    and_i_do_not_see_the_transition_call_to_action
   end
 
   scenario "There's a step by step list" do
@@ -33,6 +34,7 @@ describe "Contextual navigation" do
     and_i_visit_that_page
     then_i_see_the_step_by_step
     and_the_step_by_step_header
+    and_i_do_not_see_the_transition_call_to_action
   end
 
   scenario "There's more than one step by step" do
@@ -61,12 +63,14 @@ describe "Contextual navigation" do
     and_i_visit_that_page
     then_i_see_the_related_links_sidebar
     and_the_default_breadcrumbs
+    and_i_see_the_transition_call_to_action
   end
 
   scenario "The page has a topic" do
     given_theres_a_page_with_a_topic
     and_i_visit_that_page
     then_i_see_the_topic_breadcrumb
+    and_i_see_the_transition_call_to_action
   end
 
   scenario "The page has many topics" do
@@ -74,6 +78,7 @@ describe "Contextual navigation" do
     and_i_visit_that_page
     then_i_see_the_topic_breadcrumb
     and_i_see_the_both_topics_in_the_related_navigation_footer
+    and_i_see_the_transition_call_to_action
   end
 
   scenario "It's a HTML Publication document" do
@@ -81,6 +86,7 @@ describe "Contextual navigation" do
     and_the_page_is_an_html_publication_with_that_parent
     and_i_visit_that_page
     then_i_see_home_and_parent_links
+    and_i_see_the_transition_call_to_action
   end
 
   scenario "It's a HTML Publication with a parent with coronavirus taxon" do
@@ -89,6 +95,7 @@ describe "Contextual navigation" do
     and_i_visit_that_page
     then_i_see_home_and_parent_links
     and_i_see_the_coronavirus_contextual_breadcrumbs_for_business
+    and_i_see_the_transition_call_to_action
   end
 
   scenario "A page is tagged to the transition taxon and a step_by_step" do
@@ -97,6 +104,7 @@ describe "Contextual navigation" do
     then_i_see_the_step_by_step
     and_the_step_by_step_header
     and_i_do_not_see_the_transition_contextual_breadcrumbs
+    and_i_do_not_see_the_transition_call_to_action
   end
 
   scenario "It's a HTML Publication with a parent with breadcrumbs" do
@@ -540,7 +548,21 @@ describe "Contextual navigation" do
   def and_i_see_the_transition_related_links
     within ".gem-c-contextual-sidebar" do
       expect(page).to have_css("h2", text: "Brexit transition")
-      expect(page).to have_link("Find out what it means for you", href: "/transition")
+      expect(page).to have_link(I18n.t("components.related_navigation.transition.link_text"), href: "/transition")
+    end
+  end
+
+  def and_i_see_the_transition_call_to_action
+    within ".gem-c-contextual-sidebar" do
+      expect(page).to have_selector(".gem-c-transition-countdown")
+      expect(page).to have_css(".gem-c-transition-countdown__title", text: I18n.t("components.related_navigation.transition.title"))
+    end
+  end
+
+  def and_i_do_not_see_the_transition_call_to_action
+    within ".gem-c-contextual-sidebar" do
+      expect(page).not_to have_selector(".gem-c-transition-countdown")
+      expect(page).not_to have_css(".gem-c-transition-countdown__title", text: I18n.t("components.related_navigation.transition.title"))
     end
   end